refactor(auth): move user related data structure to the right component
