+++ backend/src/main/java/com/louisplace/backend/features/auth/auth_service/IAuthProvider.java
@@ -1,6 +1,6 @@
 package com.louisplace.backend.features.auth.auth_service;
 
-import com.louisplace.backend.features.auth.UserEntity;
+import com.louisplace.backend.features.user.UserEntity;
 
 public interface IAuthProvider {
     UserEntity authenticate(String identifier, String credential);
